refactor(native): split the PSP host into a library + thin EBOOT bin
pocketjs-psp becomes lib + bin: the mechanism modules (arena allocator
trio, QuickJS embedding, ui ffi, GE DrawList backend, pak feeder, DevTools
mailbox) and a new host module (display list, graphics init with optional
zbuffer, worker-thread boot, FPU setup, job pump, exception logging) are
now linkable by game EBOOTs; the bin keeps trace/bench/capture and the 2D
frame loop. Behavior locked: PPSSPP e2e byte-exact, tape goldens, contract.
